+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.1.1]
+
+### Added
+
+- Migrated system to Foundry v13
+
+### Fixed
+
+- Resolved compatibility issues
+- Minor bug fixes
+
 ## [0.1.0]
 
 ### Added